feat: apply also to lower level headers
