+++ packages/web-app/src/components/appli/Descriptions/Description.jsx
@@ -106,6 +106,7 @@ const Description = ({
           primary={
             <SectionTitle
               title={description.title}
+              anchorId={`description-${description.id}`}
               isDeleted={description.isDeleted}
             />
           }
