+++ packages/web-app/src/components/appli/Entry/Comments/Comment.jsx
@@ -133,6 +133,7 @@ const Comment = ({ comment, isEditAllowed, isMoving, onMoveUp, onMoveDown, isFir
             primary={
               <SectionTitle
                 title={comment.title}
+                anchorId={`comment-${comment.id}`}
                 isDeleted={comment.isDeleted}
               />
             }
