+++ packages/web-app/src/hooks/index.js
@@ -7,3 +7,4 @@ export { usePermissions } from './usePermissions';
 export { useSubscriptions } from './useSubscriptions';
 export { useUserProperties } from './useUserProperties';
 export { useMoveRelevanceWithUndo } from './useMoveRelevanceWithUndo';
+export { useAnchorScroll } from './useAnchorScroll';
